Adds Route Map overlay feature
Fixes #24 - Trip Details List and Map Overlay
Fixes #29 - The bus's route should also be projected onto the map
Fixes #30 - Hide unrelated stops on the map while a route line is being displayed

Also, checks to see if `window.google` exists, and only loads the Google Maps library if it does not. This resolves a warning in the console.

Co-authored-by: tarunsinghofficial <tarunsinghwap7@gmail.com>